Feature/stored block index alt 473 (#663)
* Refactoring for using StoredBlockIndex

* Cleanup after merge

* Add PopState to StoredBlockIndex

* Fix linux build

* Fix linux build

* Fix fuzzing

* Fix C interface

* Fix warnings

* Fix C storage

* Use StoredBlockIndex in BlockBatch interface

* Fix build

* Small fix

* Validate stored endorsements

* Cleanup

+++ include/veriblock/pop/blockchain/chain_slice.hpp
@@ -58,7 +58,7 @@ struct ChainSlice {
 
   height_t firstHeight() const { return firstHeight_; }
 
-  height_t tipHeight() const { return firstHeight() + size() - 1; }
+  height_t tipHeight() const { return (int)(firstHeight() + size()) - 1; }
 
   // for compatibility
   height_t chainHeight() const { return tip()->getHeight(); }
